[media] improve screenshot search behaviour...
ensure and make explcit that we're searching primarly for mamenames if they exist and then romname if they don't or if they returned no results. This should be the primary search order, tho it will prevent returning both mamename and romame results, but I think that's right? Also removed costly things from the loops and promoted parallel processing. Obv this logic needs echoing in the dat searches