[build] multiloader games running (messy)...
now all gametypes run on windows (still have to pass params if in romdata line, but apart from that). Example of multiloader cmd from emulators.json:

%Tool:MULTILOADER% "%SHORTROM%" "%EXEPATH%" "-L cores\mednafen_psx_hw_libretro.dll" "1"

essentially we ignore most of the features of the tool, expecting a modern emulation experience to not have to worry about 8:3 naming and disc-mounting programs. So we (sensitively - hence helper fn) take the 3rd param IFF multiloader CMD and split into elements, essentially launching the emu and game same as we would without the tool (obv this ends up in converter logic!)