[build] change emulators to dynamic load...
this is from the perspective of fresh-import: if we have no existing emulators file and we import one, the media panel didn't work until whole app reloaded. Now we dynamically load emulators (requiring dynamic loading in runGame also). However this is json-file-centric thinking: will need a rethink if swapped to a db lookup