[menu] fix devtools 'wakeup' problem...
the first IPC message to toggle devtools gets "swallowed" somehow in Electron's initialization process. It's not about timing - waiting doesn't help because the system needs that first "wakeup call" before it starts responding correctly to the toggle messages. This is a workaround to send a dummy message to wake up the system before sending the real toggle message